refactor: 영화 및 광고 아이템 목록을 만드는 메서드 도메인으로 이전
