+++ app/src/main/java/woowacourse/movie/domain/model/item/MovieListItem.kt
@@ -10,4 +10,26 @@ sealed class MovieListItem {
     data class AdItem(
         val advertisement: Advertisement,
     ) : MovieListItem()
+
+    companion object {
+        fun movieListItems(
+            movies: List<MovieItem>,
+            ads: List<AdItem>,
+        ): List<MovieListItem> {
+            val adsIterator = ads.iterator()
+            val list: List<MovieListItem> =
+                buildList {
+                    movies.forEachIndexed { index, movie ->
+                        add(movie)
+                        if (index % 3 == 2 && adsIterator.hasNext()) {
+                            add(adsIterator.next())
+                        }
+                    }
+                    while (adsIterator.hasNext()) {
+                        add(adsIterator.next())
+                    }
+                }
+            return list
+        }
+    }
 }
